Move SiteClassificationType enum out of modes module and organize MSA methods
